timer fix

+++ BackendApplication/src/main/java/com/danielvishnievskyi/backendapplication/model/dto/game/GameTimeResponseDTO.java
@@ -1,7 +1,7 @@
 package com.danielvishnievskyi.backendapplication.model.dto.game;
 
 public record GameTimeResponseDTO (
-  int whitePlayerTime,
-  int blackPlayerTime
+  long whitePlayerTime,
+  long blackPlayerTime
 ) {
 }
